move name template parts to orm
